Added converters
Added the string to role converter and the string to date converter. They are used to auto convert while adding/editing user. When add/edit user's role or date, these fields are returned as string and with the help of spring they are auto converted into role and date accordingly.